refactor(tools): remove lifecycle and provider abstractions (Phase 2A)
- Delete ToolLifecycleManager (tools don't need lifecycle management)
- Delete ToolRegistry interface (keep only concrete implementation)
- Delete monolithic ToolProviders.kt (278 lines)
- Add specific tool providers (Project, Issue, Session)
- Update DefaultToolRegistry to standalone concrete class
- Fix test configuration to include MCP tool tests
- Fix ToolRegistryTest issues (string interpolation, error codes)

Impact:
- 3 files deleted (514 lines removed)
- 3 new focused providers (126 lines added)
- Net reduction: 388 lines
- All 30 ToolRegistry tests passing
- Simplified DI configuration

Tests: 497/500 passing (99.4%)
- ToolRegistry: 30/30 tests pass ✅
- ConnectionCleanupService: 3 pre-existing failures (unrelated)
