+++ src/main/kotlin/io/spiralhouse/cycletime/infrastructure/di/MCPDependencies.kt
@@ -14,11 +14,9 @@ import io.spiralhouse.cycletime.mcp.server.MCPServerEngine
 import io.spiralhouse.cycletime.mcp.server.handlers.McpMethodHandler
 import io.spiralhouse.cycletime.mcp.server.handlers.McpMethodHandlers
 import io.spiralhouse.cycletime.mcp.tools.*
-import io.spiralhouse.cycletime.mcp.tools.providers.*
 import io.spiralhouse.cycletime.mcp.resources.ResourceProviderRegistry
 import io.spiralhouse.cycletime.mcp.resources.interfaces.ResourceRegistry
 import io.spiralhouse.cycletime.mcp.tools.DefaultToolRegistry
-import io.spiralhouse.cycletime.mcp.tools.interfaces.ToolRegistry
 import io.spiralhouse.cycletime.application.services.ProjectApplicationService
 import io.spiralhouse.cycletime.application.services.IssueApplicationService
 import io.spiralhouse.cycletime.application.services.SessionApplicationService
@@ -55,15 +53,11 @@ object MCPDependencies {
             DefaultToolRegistry()
         }
         
-        provide<ToolRegistry> {
-            resolve<DefaultToolRegistry>()
-        }
-        
         // MCP Method Handler
         provide<McpMethodHandler> {
             McpMethodHandlers(
                 resourceRegistry = resolve(),
-                toolRegistry = resolve(),
+                toolRegistry = resolve<DefaultToolRegistry>(),
                 toolInvoker = resolve<DefaultToolRegistry>()
             )
         }
@@ -75,15 +69,15 @@ object MCPDependencies {
                 protocolHandler = resolve<ProtocolHandler>(),
                 config = resolve<MCPServerConfig>(), // Resolve from DI
                 resourceRegistry = resolve<ResourceRegistry>(),
-                toolRegistry = resolve<ToolRegistry>(),
+                toolRegistry = resolve<DefaultToolRegistry>(),
                 projectResourceProvider = resolve<ProjectResourceProvider>(),
                 issueResourceProvider = resolve<IssueResourceProvider>(),
                 sessionResourceProvider = resolve<SessionResourceProvider>(),
                 workflowResourceProvider = resolve<WorkflowResourceProvider>(),
                 toolProviders = listOf(
-                    resolve<ProjectToolProvider>(),
-                    resolve<IssueToolProvider>(),
-                    resolve<SessionToolProvider>()
+                    resolve<DefaultProjectToolProvider>(),
+                    resolve<DefaultIssueToolProvider>(),
+                    resolve<DefaultSessionToolProvider>()
                 )
             )
         }
@@ -98,9 +92,9 @@ object MCPDependencies {
                     resolve<WorkflowResourceProvider>()
                 ),
                 toolProviders = listOf(
-                    resolve<ProjectToolProvider>(),
-                    resolve<IssueToolProvider>(),
-                    resolve<SessionToolProvider>()
+                    resolve<DefaultProjectToolProvider>(),
+                    resolve<DefaultIssueToolProvider>(),
+                    resolve<DefaultSessionToolProvider>()
                 )
             )
         }
@@ -134,19 +128,19 @@ object MCPDependencies {
         }
         
         // Tool Providers - connected to application services
-        provide<ProjectToolProvider> { 
+        provide<DefaultProjectToolProvider> { 
             DefaultProjectToolProvider(
                 projectService = resolve<ProjectApplicationService>()
             )
         }
         
-        provide<IssueToolProvider> { 
+        provide<DefaultIssueToolProvider> { 
             DefaultIssueToolProvider(
                 issueService = resolve<IssueApplicationService>()
             )
         }
         
-        provide<SessionToolProvider> { 
+        provide<DefaultSessionToolProvider> { 
             DefaultSessionToolProvider(
                 sessionService = resolve<SessionApplicationService>()
             )
