+++ src/main/kotlin/io/spiralhouse/cycletime/mcp/server/handlers/McpMethodHandler.kt
@@ -3,7 +3,6 @@ package io.spiralhouse.cycletime.mcp.server.handlers
 import io.spiralhouse.cycletime.mcp.protocol.JsonRpcRequest
 import io.spiralhouse.cycletime.mcp.protocol.JsonRpcResponse
 import io.spiralhouse.cycletime.mcp.protocol.JsonRpcProtocolHandler
-import io.spiralhouse.cycletime.mcp.tools.interfaces.ToolRegistry
 import io.spiralhouse.cycletime.mcp.tools.DefaultToolRegistry
 import io.spiralhouse.cycletime.mcp.resources.interfaces.ResourceRegistry
 import io.spiralhouse.cycletime.mcp.server.state.ServerState
@@ -58,7 +57,7 @@ interface McpMethodHandler {
  */
 class DefaultMcpMethodHandler(
     private val protocolHandler: JsonRpcProtocolHandler,
-    private val toolRegistry: ToolRegistry,
+    private val toolRegistry: DefaultToolRegistry,
     private val toolInvoker: DefaultToolRegistry,
     private val resourceRegistry: ResourceRegistry,
     private val serverState: ServerState = ServerState()
